released v0.8.1-next.8 - releasing as only index.ts
