--- src/StatusError.ts
@@ -1,8 +1,0 @@
-export class StatusError extends Error {
-  status: number
-
-  constructor(status = 500, message = 'Internal Error.') {
-    super(message)
-    this.status = status
-  }
-}
